requirements.txt
openai
python-dotenv
pdfplumber
fpdf 